requirements.txt
customtkinter
mysql-connector-python
pillow
python-dotenv
